+++ gdynamicDigitization/gFluxDigitization.cc
@@ -13,7 +13,7 @@ bool GFluxDigitization::defineReadoutSpecs() {
 
 
 // digitized the hit
-GDigitizedData* GFluxDigitization::digitizeHit(GHit *ghit, int hitn) {
+GDigitizedData* GFluxDigitization::digitizeHit(GHit *ghit, size_t hitn) {
 
 	// ghit->getGID() must have a single entry
 	GIdentifier identity = ghit->getGID().front();
@@ -21,7 +21,7 @@ GDigitizedData* GFluxDigitization::digitizeHit(GHit *ghit, int hitn) {
 	GDigitizedData* gdata = new GDigitizedData(ghit);
 
 	gdata->includeVariable(identity.getName(), identity.getValue()             );
-	gdata->includeVariable("hitn",             hitn                            );
+	gdata->includeVariable("hitn",             (int) hitn                      );
 	gdata->includeVariable("eTot",             ghit->getTotalEnergyDeposited() );
 	gdata->includeVariable("time",             ghit->getAverageTime()          );
